% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldpop_data.R
\name{convert_wp_agegroups}
\alias{convert_wp_agegroups}
\title{Function to transform population data from a worldpop age groups to user-defined age groups}
\usage{
convert_wp_agegroups(age_pop_data, age_groups = seq(0, 80, 5), max_age = 110)
}
\arguments{
\item{age_pop_data}{data pulled from worldpop using `load_worldpop_age`}

\item{age_groups}{lower age limit of user-defined age groups. For 5 year age groups, this would be 0, 5, 10, etc.}

\item{max_age}{max age to include (default is 100)}
}
\value{
long age population data by admin level 2, in user-defined age groups
}
\description{
Function to transform population data from a worldpop age groups to user-defined age groups
}
\references{
WorldPop (https://www.worldpop.org/geodata)
}
