% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_city_std}
\alias{get_city_std}
\title{get_city_std}
\usage{
get_city_std(location)
}
\arguments{
\item{location}{vector of city names}
}
\value{
vector of names for each location. These can then be used to identify other characteristics of the location
}
\description{
get a standardized name for a vector of Cities. This will identify most likely matched
to the inputted names used a variety of tools and methods.
}
