% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_country_code}
\alias{get_country_code}
\title{get_country_code}
\usage{
get_country_code(
  country,
  code = c("FIFA", "IOC", "ISO2", "UNcode", "MARC", "GAUL", "FIPS", "WMO", "ITU", "DS",
    "M49", "Continent", "TLD", "Geoname.ID", "EDGAR")
)
}
\arguments{
\item{country}{vector of countries}

\item{code}{vector of code types wanted (Includes: "FIFA", "IOC", "ISO2", "UNcode", "MARC", "GAUL", "FIPS", "WMO",
"ITU", "DS", "M49", "Continent", "TLD", "Geoname.ID", "EDGAR")}
}
\value{
vector of codes matching each country
}
\description{
various country codes from country names
}
