% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_country_name_ISO2}
\alias{get_country_name_ISO2}
\title{get_country_name_ISO2}
\usage{
get_country_name_ISO2(ISO2)
}
\arguments{
\item{ISO2}{vector of ISO2s}
}
\value{
vector of country names matching ISO2. These can then be used to identify other characteristics of the country
}
\description{
Standard country name from ISO2.
}
