% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_country_name_ISO3}
\alias{get_country_name_ISO3}
\title{get_country_name_ISO3}
\usage{
get_country_name_ISO3(ISO)
}
\arguments{
\item{country}{vector of ISO3s}
}
\value{
vector of country names matching each ISO. These can then be used to identify other characteristics of the country
}
\description{
Standard country name from ISO3.
}
