% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_country_name_std}
\alias{get_country_name_std}
\title{get_country_name_std}
\usage{
get_country_name_std(country)
}
\arguments{
\item{country}{vector of country names}
}
\value{
vector of country names matching each country. These can then be used to identify other characteristics of the country
}
\description{
Standard country name.
}
