% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_UNcode_from_ISO3}
\alias{get_UNcode_from_ISO3}
\title{get_UNcode_from_ISO3}
\usage{
get_UNcode_from_ISO3(ISO3)
}
\arguments{
\item{ISO3}{vector of ISO3}
}
\value{
vector of UN codes matching each ISO3
}
\description{
UN code from ISO3
}
