% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region.R
\name{get_who.region}
\alias{get_who.region}
\alias{get_whoregion}
\title{get_whoregion}
\usage{
get_whoregion(ISO)
}
\arguments{
\item{ISO}{vector of ISO3 codes}
}
\value{
vector of WHO regions
}
\description{
Identify WHO region from ISO
}
