% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_who_disease_incidence.R
\name{who_disease_incidence_year}
\alias{who_disease_incidence_year}
\title{who_disease_incidence_year}
\usage{
who_disease_incidence_year(
  disease = c("CRS", "Diphtheria", "JapEnc", "Measles", "Mumps", "Ntetanus",
    "Pertussis", "Polio", "Rubella", "Ttetanus", "Yfever"),
  save_dir = NULL,
  long_data = TRUE
)
}
\arguments{
\item{save_dir}{Where to save the xls file, if interested. If NULL, will not save.}

\item{long_data}{Logical, whether long form is desired (default).}

\item{diseaes}{Disease of interest. Options currently include "CRS","Diphtheria","JapEnc","Measles","Mumps","Ntetanus","Pertussis","Polio","Rubella","Ttetanus",and "Yfever".}
}
\value{
Tibble of reported incidence by year
}
\description{
Load monthly incidence data reported to the World Health Organization
}
