% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldpop_data.R
\name{wp_geotiff_filenames}
\alias{wp_geotiff_filenames}
\title{Function to get population data from a worldpop geotiff and geounit shapefile
This function extracts raster values for overlayed polygons}
\usage{
wp_geotiff_filenames(country = "BGD", year = "2020")
}
\arguments{
\item{country}{ISO3 of country of interest}

\item{year}{Year of population data (2000 to 2020)}
}
\value{
vector of filenames of the WorldPop age geotiffs
}
\description{
Function to get population data from a worldpop geotiff and geounit shapefile
This function extracts raster values for overlayed polygons
}
\references{
WorldPop (https://www.worldpop.org/geodata)
}
