% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_location_std}
\alias{get_location_std}
\title{get_location_std}
\usage{
get_location_std(location)
}
\arguments{
\item{location}{vector of location names}
}
\value{
vector of names for each location. These can then be used to identify other characteristics of the location
}
\description{
get a standardized name for a vector of locations. This will identify most likely matched
to the inputted names used a variety of tools and methods. This should be used when the data likely have
locations that are not only countries (i.e., some sub-countries like England or Dubai)
}
